requirements.txt
duckdb==0.10.3
click==8.1.7
jinja2==3.1.4
pandas==2.2.2
plotly==5.22.0
pydantic==2.7.4
pyyaml==6.0.1
streamlit==1.37.1
typer==0.12.3
